feat: Add dynamic background selection and prepare for obstacles
If applied, this commit will introduce a new feature that allows users to
randomly change the game background by pressing the "Change Background" button
in the pause menu. This enhancement aims to make the game more engaging for users.

Additionally, new images for obstacles have been added in preparation for the
upcoming ObstaclesModel implementation. The necessary changes have been made
to Game.FXML, GameController.java, and MenuController.java to support the
background selection functionality.

The ObstaclesModel.java class is planned for implementation
in the upcoming days to fully integrate obstacles into the game.