refactor: Generalise ItemsModel and Implement Obstacles Functionality
If applied, this commit will refactor the ItemsModel class to generalise it for different game items. The Obstacle class is removed as it's no longer needed, and instead, the ItemsModel is used for creating obstacles dynamically during the game.

The changes include:
- Refactoring of the ItemsModel class to serve as a common model for various game items, eliminating the need for a separate Obstacle class.
- Implementation of obstacle functionality in the GameController to check for collisions with the snake.
- Deduct bodyPoints from the snake and also from the score, the snake should reduce it's size as it hits the obstacles "lose a life"

Next steps involve creating an ItemsController to manage additional functionalities like slow motion, fast motion, and 2x points power-ups.